refactor: merge service-application and service-assess
