--- service-assess/.gitignore
@@ -1,31 +1,0 @@
-# Module-specific gitignore for service-assess
-
-# IntelliJ IDEA module files
-.idea/
-*.iml
-
-# RSA public key for JWT validation
-src/main/resources/rsa-public-key.pem
-src/main/resources/*.pem
-src/main/resources/*.key
-
-# Local configuration files
-src/main/resources/application-local.yml
-src/main/resources/application-dev.yml
-src/main/resources/application-secret.yml
-
-# Module build artifacts
-build/
-target/
-out/
-
-# Temporary files
-*.tmp
-*.temp
-
-# AI/ML models and data (if any)
-*.model
-*.pkl
-*.h5
-models/
-datasets/
