--- service-assess/build.gradle
@@ -1,16 +1,0 @@
-// service-assess specific dependencies
-dependencies {
-    // Shared persistence module
-    implementation project(':persistence')
-    
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.springframework.boot:spring-boot-starter-security'
-    implementation 'org.springframework.boot:spring-boot-starter-oauth2-resource-server'
-    
-    // Database
-    runtimeOnly 'org.postgresql:postgresql'
-    
-    // Testing
-    testImplementation 'org.springframework.security:spring-security-test'
-    testImplementation 'org.testcontainers:postgresql'
-}
