--- service-assess/src/main/resources/application.yml
@@ -1,42 +1,0 @@
-server:
-  port: ${SERVER_PORT:8083}
-
-spring:
-  application:
-    name: assess-service
-
-  datasource:
-    url: jdbc:postgresql://${DB_HOST:localhost}:${DB_PORT:5432}/hrapp?stringtype=unspecified
-    username: ${DB_USERNAME:postgres}
-    password: ${DB_PASSWORD:postgres}
-    driver-class-name: org.postgresql.Driver
-
-  jpa:
-    hibernate:
-      ddl-auto: validate
-    show-sql: true
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.PostgreSQLDialect
-        format_sql: true
-
-# JWT Configuration for token validation
-app:
-  jwt:
-    public-key: ${JWT_PUBLIC_KEY}
-
-# Logging configuration
-logging:
-  level:
-    de.tum.devops.assess: DEBUG
-    org.springframework.security: DEBUG
-    org.hibernate.SQL: DEBUG
-  pattern:
-    console: "%d{yyyy-MM-dd HH:mm:ss} - %msg%n"
-
-# CORS configuration
-cors:
-  allowed-origins: ${CORS_ALLOWED_ORIGINS:"http://localhost:3000,http://localhost:5173,http://localhost:4200"}
-  allowed-methods: GET,POST,PUT,DELETE,OPTIONS
-  allowed-headers: "*"
-  allow-credentials: true
