+++ settings.gradle
@@ -6,14 +6,12 @@ include 'persistence'
 include 'service-auth'
 include 'service-job'
 include 'service-application'
-include 'service-assess'
 
 // Set project directories
 project(':persistence').projectDir = file('persistence')
 project(':service-auth').projectDir = file('service-auth')
 project(':service-job').projectDir = file('service-job')
 project(':service-application').projectDir = file('service-application')
-project(':service-assess').projectDir = file('service-assess')
 
 // Enable Gradle features
 enableFeaturePreview('TYPESAFE_PROJECT_ACCESSORS')
